These are all the MATLAB m-files and scripts
The m-files are both production and deprecated versions, uploaded without differentiation.  Some were used as standalone files, some were used in larger MATLAB command line programs, and some are interchangeable.  Some were also used in the modeling of polymer chains, which is stored more explicitly in a separate repository on my account.  It is possible that (or these) have inter-dependencies that prevent functioning without the other files.  There may also be redundancies and copies in both repositories, though I attempted to minimize that.  This also contains script files that are merely long lists of commands to analyze data.  That data is not here, but the scripts can provide some information on how to properly call the m-files as well as the bulk analysis capabilities that I never fully explored.